Some more patches concerning arxan and DW
Need to fix DW_Frame to kill reserved hardware breakpoints
